nethist: Network Histograms

Estimates network histograms, a blockmodel approximation to the graphon underlying a network's connectivity pattern, for both single-layer and multilayer networks. Implements graphon estimation methods including the profile-likelihood method of Olhede and Wolfe (2014) doi:10.1073/pnas.1400374111 and the least-squares method of Gao, Lu, and Zhou (2015) doi:10.1214/15-AOS1354 for single-layer networks, and the multilayer extension of Song and Olhede (2026) arXiv:2608.14536.

See also

Author

Maintainer: Youngseok Song youngseok.song@mail.wvu.edu

Authors:

Youngseok Song youngseok.song@mail.wvu.edu

Sofia C. Olhede sofia.olhede@epfl.ch